[✨] Fix containers to allow external queries and manipulations.
📑 Description:
In fault of external queries and manipulations
(guards/result object this situation )
We need to change system of pointer using shared_ptr and
weak_ptr to give to these queries results.

ℹ Why is it necessary ?
- templates can't handle pointer type

🔝 How does the change address the issue ?
- Using shared_ptr inside domain object and create weak ptr to give to
query result.

⚠ What side effects does this change have ?
- Change sementic.

📅TODO for the future:
- Maybe change shared to copy of unique_ptr and generation of simple ptr
result.
If we do that, we need to create an overhead of the result object to
works with classic pointer.

Note:
Change this system.